ground work for exposing traits from Rust
