+++ mica-language/src/value/traits.rs
@@ -1,5 +1,9 @@
-use crate::bytecode::{DispatchTable, Opr24};
-use crate::gc::GcRaw;
+use std::rc::Rc;
+
+use crate::bytecode::{DispatchTable, Environment, Opr24};
+use crate::gc::{GcRaw, Memory};
+
+use super::Closure;
 
 /// Instance of a trait.
 pub struct Trait {
@@ -12,4 +16,24 @@ impl Trait {
       // SAFETY: This operation is safe, because the dtable of a trait can never be modified.
       unsafe { self.dtable.get() }
    }
+}
+
+/// Creates a new instance of a trait inside the given GC memory.
+pub fn create_trait(env: &mut Environment, gc: &mut Memory, trait_id: Opr24) -> GcRaw<Trait> {
+   let prototype = env.get_trait(trait_id).expect("trait with given ID does not exist");
+   let name = &prototype.name;
+   let mut dispatch_table = DispatchTable::new_for_type(Rc::clone(name));
+   dispatch_table.pretty_name = Rc::from(format!("trait {name}"));
+   for &(method_id, function_id) in &prototype.shims {
+      let closure = gc.allocate(Closure {
+         function_id,
+         captures: vec![],
+      });
+      dispatch_table.set_method(method_id, closure);
+   }
+   let dispatch_table = gc.allocate(dispatch_table);
+   gc.allocate(Trait {
+      id: trait_id,
+      dtable: dispatch_table,
+   })
 }
